fix calculate date and FE process detail

+++ Fall2023_SWP391_NET1704_Vitvang/src/main/java/com/vitvang/productionmanagement/controller/DeleteMaterialofCageController.java
@@ -4,16 +4,20 @@
  */
 package com.vitvang.productionmanagement.controller;
 
+import com.vitvang.productionmanagement.dao.material.MaterialDAO;
+import com.vitvang.productionmanagement.model.UserDTO;
+import com.vitvang.productionmanagement.util.Constant;
+import static com.vitvang.productionmanagement.util.tool.checkRole;
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import java.io.IOException;
 import java.sql.SQLException;
 import javax.naming.NamingException;
-import com.vitvang.productionmanagement.dao.material.MaterialDAO;
 
 /**
  *
@@ -32,7 +36,16 @@ public class DeleteMaterialofCageController extends HttpServlet {
             String CageID = request.getParameter("txtCageID").trim();
             String url = ERROR_PAGE;
             try {
-
+                  HttpSession session = request.getSession();// phai luon co san session
+                  UserDTO currUser = (UserDTO) session.getAttribute("USER");
+                  if (currUser == null) {
+                        return;
+                  }
+                  int roleID = currUser.getRoleID();
+                  //0. check role 
+                  if (!checkRole(roleID, Constant.isManager) && !checkRole(roleID, Constant.isStaff)) {
+                        return;
+                  }
                   MaterialDAO dao = new MaterialDAO();
                   dao.deleteMaterialOfCage(materialID, CageID);
                   url = "MainController"
